requirements.txt
pulsar-client
boto3